requirements.txt
accelerate==1.10.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiolimiter==1.2.1
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
attrs==25.3.0
autoflake==2.3.1
azure-ai-documentintelligence==1.0.2
azure-core==1.35.0
azure-identity==1.24.0
babel==2.17.0
backrefs==5.9
beautifulsoup4==4.13.4
black==25.1.0
blis==1.3.0
bracex==2.6
build==1.3.0
cachetools==5.5.2
catalogue==2.0.10
certifi==2025.8.3
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.4.3
click==8.2.1
cloudpathlib==0.21.1
cobble==0.1.4
colorama==0.4.6
coloredlogs==15.0.1
confection==0.1.5
coverage==7.10.3
cramjam==2.11.0
cryptography==45.0.6
cssselect2==0.8.0
cymem==2.0.11
defusedxml==0.7.1
dill==0.4.0
distlib==0.4.0
distro==1.9.0
docling==2.44.0
docling-core==2.44.2
docling-ibm-models==3.9.0
docling-parse==4.1.0
docutils==0.22
easyocr==1.7.2
et-xmlfile==2.0.0
fastjsonschema==2.21.2
fastparquet==2024.11.0
ffmpeg==1.4
ffmpeg-downloader==0.4.0
filelock==3.19.1
filetype==1.2.0
flake8==7.3.0
flatbuffers==25.2.10
frozenlist==1.7.0
fsspec==2025.7.0
ghp-import==2.1.0
gitdb==4.0.12
gitpython==3.1.45
google-auth==2.40.3
google-genai==1.32.0
griffe==1.12.1
h11==0.16.0
hf-xet==1.1.7
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.34.4
humanfriendly==10.0
id==1.5.0
identify==2.6.13
idna==3.10
imageio==2.37.0
iniconfig==2.1.0
isodate==0.7.2
isort==6.0.1
jaraco-classes==3.4.0
jaraco-context==6.0.1
jaraco-functools==4.2.1
jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonlines==3.1.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
jupyter-core==5.8.1
keyring==25.6.0
langchain-core==0.3.74
langchain-text-splitters==0.3.9
langcodes==3.5.0
langsmith==0.4.14
language-data==1.3.0
latex2mathml==3.78.0
lazy-loader==0.4
lxml==5.4.0
magika==0.6.2
mammoth==1.10.0
marisa-trie==1.2.1
markdown==3.8.2
markdown-it-py==4.0.0
markdownify==1.2.0
markitdown==0.1.2
marko==2.2.0
markupsafe==3.0.2
mccabe==0.7.0
mdurl==0.1.2
mergedeep==1.3.4
mkdocs==1.6.1
mkdocs-autorefs==1.4.2
mkdocs-awesome-pages-plugin==2.10.1
mkdocs-get-deps==0.2.0
mkdocs-glightbox==0.4.0
mkdocs-material==9.6.17
mkdocs-material-extensions==1.3.1
mkdocstrings==0.30.0
mkdocstrings-python==1.17.0
more-itertools==10.7.0
mpire==2.10.2
mpmath==1.3.0
msal==1.33.0
msal-extensions==1.3.1
multidict==6.6.4
multiprocess==0.70.18
murmurhash==1.0.13
mypy-extensions==1.1.0
natsort==8.4.0
nbformat==5.10.4
networkx==3.5
nh3==0.3.0
ninja==1.13.0
nltk==3.9.1
nodeenv==1.9.1
numpy==2.3.2
olefile==0.47
onnxruntime==1.22.1
openai==1.99.9
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
orjson==3.11.2
packaging==25.0
paginate==0.5.7
pandas==2.3.1
pastel==0.2.1
pathspec==0.12.1
pdfminer-six==20250506
pdfplumber==0.11.7
pillow==11.3.0
platformdirs==4.3.8
pluggy==1.6.0
poethepoet==0.37.0
pre-commit==4.3.0
preshed==3.0.10
propcache==0.3.2
protobuf==5.29.5
psutil==7.0.0
pyasn1==0.6.1
pyasn1-modules==0.4.2
pyclipper==1.3.0.post6
pycodestyle==2.14.0
pycparser==2.22
pydantic==2.11.7
pydantic-core==2.33.2
pydantic-settings==2.10.1
pydub==0.25.1
pyflakes==3.4.0
pygments==2.19.2
pyjwt==2.10.1
pylatexenc==2.10
pymdown-extensions==10.16.1
pymupdf==1.26.3
pypandoc==1.15
pypdf==6.0.0
pypdfium2==4.30.0
pyproject-autoflake==1.0.2
pyproject-hooks==1.2.0
pytest==8.4.1
pytest-cov==6.2.1
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
python-pptx==1.0.2
pytz==2025.2
pyupgrade==3.20.0
pyyaml==6.0.2
pyyaml-env-tag==1.1
readme-renderer==44.0
referencing==0.36.2
regex==2025.7.34
reportlab==4.4.3
requests==2.32.4
requests-mock==1.12.1
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==14.1.0
rpds-py==0.27.0
rsa==4.9.1
rtree==1.4.1
ruff==0.12.9
safetensors==0.6.2
scikit-image==0.25.2
scikit-learn==1.7.1
scipy==1.16.1
semchunk==2.2.2
sentence-transformers==5.1.0
setuptools==80.9.0
setuptools-scm==8.3.1
shapely==2.1.1
shellingham==1.5.4
six==1.17.0
smart-open==7.3.0.post1
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
speechrecognition==3.14.3
srsly==2.5.1
svglib==1.5.1
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
thinc==8.3.6
threadpoolctl==3.6.0
tifffile==2025.6.11
tiktoken==0.11.0
tinycss2==1.4.0
tokenize-rt==6.2.0
tokenizers==0.21.4
toml==0.10.2
torch==2.7.1
torchvision==0.22.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.55.2
twine==6.1.0
typer==0.16.0
typing-extensions==4.14.1
typing-inspection==0.4.1
tzdata==2025.2
urllib3==2.5.0
virtualenv==20.34.0
voyageai==0.3.4
wasabi==1.1.3
watchdog==6.0.0
wcmatch==10.1
weasel==0.4.1
webencodings==0.5.1
websockets==15.0.1
wheel==0.45.1
wrapt==1.17.3
xlrd==2.0.2
xlsxwriter==3.2.5
yarl==1.20.1
youtube-transcript-api==1.0.3
zstandard==0.23.0
